Script to web-scrape the portal Habitaclia
This script scrapes the desired number of result pages of the portal Habitaclia, for the city of Barcelona. Returns an Excel file with the complete information and a backup shelve file with the location-distance relations related to the address. The variable "DistanceToCenter" is calculated in kilometers with respect to Plaça Sant Jaume, in straight line.