Script to web-scrape the portal Pisos
This script scrapes the desired number of result pages of the portal Pisos, for the city of Barcelona. Returns an Excel file with the complete information; a backup shelve file with the location-distance relations related to the address; and a backup shelve with the links to each ad. The variable "DistanceToCenter" is calculated in kilometers with respect to Plaça Sant Jaume, in straight line. The user is given the choice to harvest the urls anew or use a backup shelve from the same day.

URL to the portal: https://www.pisos.com/